backend/requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
sqlalchemy
psycopg2-binary
python-dotenv
